Refactor the project structure to enhance readability and maintainability
