--- dnn_from_scratch/src/dnn/mod.rs
@@ -1,2 +1,0 @@
-pub mod fully_connected;
-pub mod neural_network;
