+++ dnn_from_scratch/src/lib.rs
@@ -3,9 +3,9 @@ extern crate ndarray as nd;
 extern crate rand;
 extern crate rand_distr;
 
-pub mod dnn;
-pub mod utils;
-
-pub use crate::dnn::fully_connected;
-pub use crate::dnn::neural_network;
-pub use crate::utils::report;
+mod activation;
+pub mod classification;
+mod fully_connected;
+pub mod loss;
+pub mod neural_network;
+pub mod report;
