+++ dnn_from_scratch/src/report.rs
