+++ src/mnist_experiment/mod.rs
@@ -1,4 +1,6 @@
 extern crate dnn_from_scratch;
+use dnn_from_scratch::classification;
+use dnn_from_scratch::loss::Loss;
 use dnn_from_scratch::neural_network::NeuralNetwork;
 use dnn_from_scratch::report::ReportData;
 use nd::Array2;
@@ -31,25 +33,30 @@ impl Training for NeuralNetwork {
     ) {
         let mut learning_rate;
         let mut report_data = ReportData::new(n_epochs, "accuracy");
+        let loss = Loss::new("cross_entropy");
         for epoch in 1..=n_epochs {
             // Training pipeline
             let output = self.forward(&x_train);
-            let train_loss = self.categorical_cross_entropy(&output, &y_train);
-            let predicted_labels = self.argmax(&output, 1);
-            let true_labels = self.argmax(&y_train, 1);
-            let train_accuracy =
-                self.compute_accuracy(&predicted_labels.into_dyn(), &true_labels.into_dyn());
+            let train_loss = loss.compute_loss(&output, &y_train);
+            let predicted_labels = classification::argmax(&output, 1);
+            let true_labels = classification::argmax(&y_train, 1);
+            let train_accuracy = classification::compute_accuracy(
+                &predicted_labels.into_dyn(),
+                &true_labels.into_dyn(),
+            );
             let scaling_factor = 6. / output.shape()[0] as f64;
             let output_gradient = (output - y_train.clone()).map(|&v| v * scaling_factor);
             learning_rate = initial_learning_rate / (1. + decay * epoch as f64);
             self.backward(&output_gradient, learning_rate, epoch);
             // Testing pipeline
             let output = self.forward(&x_test);
-            let test_loss = self.categorical_cross_entropy(&output, &y_test);
-            let predicted_labels = self.argmax(&output, 1);
-            let true_labels = self.argmax(&y_test, 1);
-            let test_accuracy =
-                self.compute_accuracy(&predicted_labels.into_dyn(), &true_labels.into_dyn());
+            let test_loss = loss.compute_loss(&output, &y_test);
+            let predicted_labels = classification::argmax(&output, 1);
+            let true_labels = classification::argmax(&y_test, 1);
+            let test_accuracy = classification::compute_accuracy(
+                &predicted_labels.into_dyn(),
+                &true_labels.into_dyn(),
+            );
             // Report
             report_data.add(train_loss, train_accuracy, test_loss, test_accuracy);
             report_data.print_report(epoch);
