+++ src/rssi_experiment/mod.rs
@@ -1,3 +1,4 @@
+use dnn_from_scratch::loss::Loss;
 use dnn_from_scratch::neural_network::NeuralNetwork;
 use dnn_from_scratch::report::ReportData;
 use nd::Array2;
@@ -31,19 +32,21 @@ impl Training for NeuralNetwork {
     ) {
         let mut learning_rate;
         let mut report_data = ReportData::new(n_epochs, "error");
+        let loss = Loss::new("mse");
+        let error = Loss::new("rmse");
         for epoch in 1..=n_epochs {
             // Training pipeline
             let output = self.forward(&x_train);
-            let train_loss = self.mse(&output, &y_train);
-            let train_accuracy = self.rmse(&output, &y_train);
+            let train_loss = loss.mse(&output, &y_train);
+            let train_accuracy = error.rmse(&output, &y_train);
             let scaling_factor = 1. / output.shape()[0] as f64;
             let output_gradient = (output - y_train.clone()).map(|&v| v * scaling_factor);
             learning_rate = initial_learning_rate / (1. + decay * epoch as f64);
             self.backward(&output_gradient, learning_rate, epoch);
             // Testing pipeline
             let output = self.forward(&x_test);
-            let test_loss = self.mse(&output, &y_test);
-            let test_accuracy = self.rmse(&output, &y_test);
+            let test_loss = loss.mse(&output, &y_test);
+            let test_accuracy = error.rmse(&output, &y_test);
             // Report
             report_data.add(train_loss, train_accuracy, test_loss, test_accuracy);
             report_data.print_report(epoch);
